created Constants for list size, and functions to generate that many list

+++ tour.cpp
@@ -1,7 +1,7 @@
 #include <cmath>
 #include "tour.hpp"
 
-tour::tour(list<city> cities, int rating) {
+tour::tour(list<city> cities, double rating) {
     locations = cities;
     fitness = rating;
 }
